Pull XCDevice out of xcode.dart
