+++ packages/flutter_tools/lib/src/ios/devices.dart
@@ -22,7 +22,7 @@ import '../convert.dart';
 import '../device.dart';
 import '../device_port_forwarder.dart';
 import '../globals.dart' as globals;
-import '../macos/xcode.dart';
+import '../macos/xcdevice.dart';
 import '../project.dart';
 import '../protocol_discovery.dart';
 import '../vmservice.dart';
